refactor(main): remove main service, error handling, and version modules, consolidating main execution and project settings
