+++ src/declusor/__init__.py
@@ -1,14 +1,10 @@
 from . import command, config, controller, core, interface, util
-from .version import PROJECT_DESCRIPTION, PROJECT_NAME, __version__
 
 __all__ = [
-    "__version__",
     "command",
     "config",
     "controller",
     "core",
     "interface",
-    "PROJECT_DESCRIPTION",
-    "PROJECT_NAME",
     "util",
 ]
