--- src/declusor/version.py
@@ -1,6 +1,0 @@
-major, minor, patch = 0, 2, 0
-
-__version__ = f"{major}.{minor}.{patch}"
-
-PROJECT_NAME = "declusor"
-PROJECT_DESCRIPTION = "a versatile tool for delivering Bash payloads to Linux systems."
